+++ src/index.js
@@ -1,3 +1,7 @@
+// Elements
 export { default as Button } from './components/Elements/Button';
 export { default as Icon } from './components/Elements/Icon';
 export { default as Spinner } from './components/Elements/Spinner';
+
+// Forms
+export { default as Toggle } from './components/Forms/Toggle';