+++ src/shared/index.js
@@ -2,6 +2,7 @@
 export { default as BaseProps } from './types/BaseProps';
 export { default as Color } from './types/Color';
 export { default as Size } from './types/Size';
+export { default as BaseFormProps } from './types/BaseFormProps';
 
 // functions
 export { default as makeCallback } from './functions/makeCallback';