+++ src/shared/types/BaseFormProps.js
@@ -1,0 +1,17 @@
+import PropTypes from 'prop-types';
+
+export default {
+  /**
+   * The name of the input field
+   * @default: undefined
+   */
+  name: PropTypes.string,
+  /**
+   * The input label text
+   */
+  label: PropTypes.string,
+  /**
+   * The input help text
+   */
+  helpText: PropTypes.string,
+};