+++ src/shared/types/BaseProps.js
@@ -2,15 +2,15 @@ import PropTypes from 'prop-types';
 
 export default {
   /**
-   * the id to be applied to the component's element
+   * the id to be applied to the component's first element
    */
   id: PropTypes.string,
   /**
-   * the CSS classes to be merged within the component's classes
+   * the CSS classes to be merged with the component's classes
    */
   className: PropTypes.string,
   /**
-   * the CSS rules to be applied to the component's element
+   * the CSS rules to be applied to the component's first element
    */
   style: PropTypes.object,
 };